+++ inst/shiny_apps/radial_tour/ui.R
@@ -1,36 +1,27 @@
-### "Intro" ui.R -----
+### "radial_tour" ui.R -----
+require("spinifex")
+require("shinythemes")     ## Themes for shiny, think css files.
+require("shinycssloaders") ## shinycssloaders::withSpinner()
 
-##### Setup ----
-.include_dev_display  <- FALSE
-.include_obs_msg      <- FALSE
-.obs_msg_counter <- 0L
-## Creat lontextLine, a string containing App name, spinifex version, and sys date.
+##### Initialize ----
+## Create contextLine, a string containing App name, spinifex version, and sys date.
 .wd <- getwd()
 .regex <- regexpr("\\/[^\\/]*$", .wd)
 .local_path <- substr(.wd, .regex + 1L, nchar(.wd))
-contextLine <- paste0("Spinifex app, '", .local_path, 
+contextLine <- paste0("Spinifex app, '", .local_path,
                       "' --- (spinifex v", packageVersion("spinifex"),
                       ") --- Ran on ", Sys.Date())
 
-require("spinifex")
-require("tourr")
-require("ggplot2")
-require("tibble")
-require("shinythemes") ## Themes for shiny, think css files.
-require("shinyjs")     ## Extend JavaScript (Think HTML interactivity) control and formating, 
-## also see ?shinyjs::toggle   &   https://daattali.com/shiny/shinyjs-basic/
-
-default_projVars <- names(tourr::flea[, 1L:6L])
-
-### Data Tab -----
+### tabData -----
 tabData <- tabPanel(
-  "Data", fluidPage(
+  "Process projection data", 
+  fluidPage(
     sidebarPanel(
       width = 3L,
       ## Select data
       selectInput("dat", "Dataset",
-                  c("flea", "olive", "weather", "wine", "breast cancer",
-                    "diabetes, long", "diabetes, wide"), #Upload file"), ## upload causing other errors
+                  c("flea", "Upload file", "olive", "weather", "wine", "breast cancer",
+                    "diabetes, long", "diabetes, wide"),
                   "flea"),
       conditionalPanel("input.dat == 'Upload file'",
                        fileInput("data_file", "Data file",
@@ -39,27 +30,24 @@ tabData <- tabPanel(
                        ),
                        verbatimTextOutput("data_msg")),
       ## Select the projection variables
-      checkboxGroupInput(
-        "projVars",
-        label = "Projection variables",
-        choices  = default_projVars,
-        selected = default_projVars
-      ),
+      uiOutput("inputProjVars"),
+      p("Rows with NA values excluded if present."),
       checkboxInput("rescale_data", "Standardize values to by Std.Dev.", value = TRUE)
     ),
     mainPanel(width = 9L,
-              h4("Raw input data summary"),
+              h4("Input data summary"),
               verbatimTextOutput("rawDat_summary"),
-              h4("Selected data summary"),
+              h4("Processed data summary"),
               verbatimTextOutput("selDat_summary")
     )
     
   )
 )
 
-### Radial Manual Tab ----
+### tabRadial ----
 tabRadial <- tabPanel(
-  "Radial manual tour", fluidPage(
+  "Radial manual tour", 
+  fluidPage(
     sidebarPanel(
       width = 3L,
       selectInput('manip_var_nm', 'Manip var', 1L),
@@ -68,30 +56,22 @@ tabRadial <- tabPanel(
     ),
     mainPanel(
       width = 9L,
-      plotly::plotlyOutput("plotlyAnim",
-                           height = "600px")
+      shinycssloaders::withSpinner(
+        plotly::plotlyOutput("plotlyAnim", height = "600px"), 
+        type = 8L)
     )
   )
-)
+) ## Assign tab
 
-##### ui -----
-### Tabs combined
+##### ui, combine tabs -----
 ui <- fluidPage(theme = shinythemes::shinytheme("flatly"), 
                 ## Esp see the themes: "flatly", "spacelab", "journal"
                 ## Make the lines, hr() black:
                 tags$head(tags$style(HTML("hr {border-top: 1px solid #000000;}"))),
-                shinyjs::useShinyjs(),
                 #### Content:
                 navbarPage(paste0("Spinifex app -- ", .local_path, ""),
                            tabData,
                            tabRadial
                 ),
-                h5(contextLine, style = "color: #A9A9A9"),
-                shinyjs::hidden(
-                  div(id = "dev_toggle",
-                      actionButton("browser", "browser()"),
-                      verbatimTextOutput("dev_msg")
-                  )
-                )
 )